+++ SourceCode/Services/Implementations/LanguageUtility.cs
@@ -36,7 +36,9 @@ public static class LanguageUtility
     /// Returns a list of <see cref="CultureInfo"/> that are fully supported in user interface.
     /// </summary>
     public static IList<CultureInfo> FullySupportedCultures => LanguageCultureMap.Values.Take(5).ToArray();
+    public static IList<CultureInfo> CargoNameSupportedCultures => LanguageCultureMap.Values.Take(9).ToArray();
     public static string[] FullySupportedLanguages => FullySupportedCultures.Select(c => c.TwoLetterISOLanguageName).ToArray();
+    public static string[] CargoSupportedLanguages => CargoNameSupportedCultures.Select(c => c.TwoLetterISOLanguageName).ToArray();
 
     public static IEnumerable<CultureInfo> AllSupportedCultures => LanguageCultureMap.Values;
 
@@ -45,16 +47,18 @@ public static class LanguageUtility
     /// </summary>
     private static IDictionary<Language, CultureInfo> LanguageCultureMap =>
          new Dictionary<Language, CultureInfo>() {
+                 // User interface is supported
                  { Language.English, new CultureInfo("en-GB") },
                  { Language.German, new CultureInfo("de-DE") },
                  { Language.Danish, new CultureInfo("da-DK") },
                  { Language.Swedish, new CultureInfo("sv-SE") },
                  { Language.Norwegian, new CultureInfo("nb-NO") },
-                 // Not fully supported below:
+                 // Cargo names are supported
                  { Language.Dutch, new CultureInfo("nl-NL") },
                  { Language.Polish, new CultureInfo("pl-PL") },
                  { Language.Italian, new CultureInfo("it-IT") },
                  { Language.French, new CultureInfo("fr-FR") },
+                 // Only country and flag supported
                  { Language.Czech, new CultureInfo("cs-CZ") },
                  { Language.Slovak, new CultureInfo("sk-SK") },
                  { Language.Hungarian, new CultureInfo("hu-HU") },
@@ -67,8 +71,7 @@ public static class LanguageUtility
         "sv" => "se",
         "da" => "dk",
         "nb" => "no",
-        "de" => "de",
-        _ => ""
+        _ => culture.TwoLetterISOLanguageName
     };
 
     public static CultureInfo FullySupportedOrDefaultCulture(this string? twoLetterISOLanguageName) =>