+++ RELEASENOTES.md
@@ -10,6 +10,7 @@ Release date 2023-10-09
 - **Bug fix** of user status now showing all possible statuses.
 - **Bug fix** of page history.
 - **Fix** of permitted characters and capitalisation in station customer name. 
+- **Fix** of language selection in waybills for unsupported languages for cargo names.
 - **Time-out** improvements of user interface.
 - **Delete**-button moved for *meetings* and *persons* (affects administrators).
 ## Release 1.9.5